# Quiz-Engine-Service/src/main/java/com/dhiraj/quizengineservice/dto/QuizResponse.java
package com.dhiraj.quizengineservice.dto;


import lombok.Builder;
import lombok.Data;

@Data
@Builder
public class QuizResponse {
    private String id;
    private String title;
    private String category;
    private int totalQuestions;
}