# Quiz-Engine-Service/src/main/java/com/dhiraj/quizengineservice/model/Response.java
package com.dhiraj.quizengineservice.model;

import lombok.Data;
import lombok.RequiredArgsConstructor;

@Data
@RequiredArgsConstructor
public class Response {
    private String id;
    private String option;
}